Refactor de la estructura del proyecto eliminando entidades e interfaces obsoletas, actualizando las referencias al servicio de logging y agregando una nueva carpeta para las interfaces.
