+++ Core/Enums/ProcessingStatus.cs
@@ -1,0 +1,11 @@
+﻿
+namespace Core.Enums
+{
+   public enum ProcessingStatus
+    {
+        Pending,
+        Processing,
+        Completed,
+        Failed
+    }
+}
