+++ Core/Interfaces/ICache.cs
@@ -1,0 +1,12 @@
+using Core.Entities;
+
+namespace Core.Interfaces
+{
+    public interface ICache
+    {
+        Document Get(Guid id);
+        void Set(Guid id, Document document);
+        void Remove(Guid id);
+        void Clear();
+    }
+}
