+++ Core/Interfaces/IDocumentProcessor.cs
@@ -1,0 +1,11 @@
+using Core.Entities;
+using Core.Enums;
+
+namespace Core.Interfaces
+{
+    public interface IDocumentProcessor
+    {
+        void ProcessDocument(Document document);
+        ProcessingStatus CheckStatus(Guid documentId);
+    }
+}
