+++ Core/Interfaces/IDocumentStorage.cs
@@ -1,0 +1,14 @@
+using Core.Entities;
+using System.Collections.Generic;
+
+namespace Core.Interfaces
+{
+    public interface IDocumentStorage
+    {
+        Document GetById(Guid id);
+        List<Document> GetAll();
+        Document Save(Document document);
+        void Update(Document document);
+        void Delete(Guid id);
+    }
+}
