+++ IACHRSeniorConsultantInterview/Program.cs
@@ -1,11 +1,5 @@
 using Infrastructure;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using System;
-using System.IO;
 using System.Text.Json;
-using Microsoft.Extensions.Logging;
 
 var builder = WebApplication.CreateBuilder(args);
